third_party/WebKit/LayoutTests/external/wpt/css/css-masking/parsing/mask-size-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask-size with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-mask-size">
<meta name="assert" content="mask-size supports the full grammar '<bg-size>#'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("mask-size", "cover");
test_valid_value("mask-size", "contain");

test_valid_value("mask-size", "0", "0px");
test_valid_value("mask-size", "10px");
test_valid_value("mask-size", "20%");
test_valid_value("mask-size", "auto");
test_valid_value("mask-size", "10px auto");

test_valid_value("mask-size", "10px 20%, cover, contain, auto");
</script>
</body>
</html>
